refactor(nvim): move all language server configs to their own folder
