+++ nvim/.config/nvim/lua/aivanoved/lazy/lsp/language_servers/init.lua
@@ -1,0 +1,1 @@
+return {}
