+++ nvim/.config/nvim/lua/aivanoved/lazy/lsp/language_servers/typos_lsp.lua
